+++ src/Exception/InvalidEligibilityInputException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Maatify\Eligibility\Exception;
+
+use Maatify\Exceptions\Exception\Validation\InvalidArgumentMaatifyException;
+
+final class InvalidEligibilityInputException extends InvalidArgumentMaatifyException implements EligibilityExceptionInterface
+{
+}